WiFi scanning crashing in Interactive Mode was the result of recent changes in anticipation of implementing 5GHz WiFi channels. I incorrectly thought that would require uint16_t channels because I can't count, and had updated channels[] accordingly - except I'd only partially updated channel variables, some expected uint8_t and others uint16_t. I'm unsure why the issue only appeared with Interactive Mode enabled, but it's fixed now. Channel initialisation has now (subsequent to fixing the issue) been implemented with more rigour, and only performed if scanning is started without specified channels. Starting to look good now - AP & STA identification, channel hopping both working well. AP/STA association looks good in ESP32 from limited testing (Macca's carpark), hopefully Flipper will be nice and stable now and I can iron out any remaining issues with the device list scene...

+++ Flipper/wendigo_common_defs.h
@@ -130,7 +130,7 @@ typedef struct wendigo_wifi_ap {
     void **stations;                      /** wendigo_device** */
     uint8_t stations_count;               /** Count of devices in stations */
     uint8_t ssid[MAX_SSID_LEN + 1];       /** SSID of AP */
-    uint16_t channel;
+    uint8_t channel;
     wifi_auth_mode_t authmode;
     uint32_t phy_11b: 1;                  /**< Bit: 0 flag to identify if 11b mode is enabled or not */
     uint32_t phy_11g: 1;                  /**< Bit: 1 flag to identify if 11g mode is enabled or not */
@@ -148,7 +148,7 @@ typedef struct wendigo_wifi_ap {
 typedef struct wendigo_wifi_sta {
     void *ap;                   /* wendigo_device* */
     uint8_t apMac[MAC_BYTES];
-    uint16_t channel;
+    uint8_t channel;
 } wendigo_wifi_sta;
 
 typedef struct wendigo_device {
